solutions report: rename pages
