+++ app/layout/scenarios/reports/solutions/selected-solution-page/index.ts
