+++ app/layout/scenarios/reports/solutions/selected-solution-page/map/component.tsx
