+++ app/layout/scenarios/reports/solutions/selected-solution-page/map/index.ts
