+++ app/layout/scenarios/reports/solutions/settings-page/constants.ts
