+++ app/layout/scenarios/reports/solutions/settings-page/index.ts
