+++ app/layout/scenarios/reports/solutions/solution-frequency-page/index.ts
