+++ app/layout/scenarios/reports/solutions/solution-frequency-page/map/component.tsx
